Use hashed input args in cache key (not it's text directly)
Input args are controlled by the user, and could potentially contain
anything. A common occurence is for example `--only group1,group2`. In
this example, the args contain a comma character, which is however not
allowed to be present in a cache key.

To solve this issue, rather than using the input arguments directly as a
part of the cache key, construct a sha256 hash of them and use it
instead.

This hash should not contain any prohibited characters, as it's output
is in base64 (0-9, a-f), all of which are allowed to be a part of the
cache key.